Input simulation
Simulation of the total radiation created by several users when entering
randomly in a map of some given dimensions and video plotting. The
radiation of each user is modeled as gaussian. The users can enter in a
random time, and in a random position along the boundaries of the map.
When a user cross a limit, its path is finish.