RF model: ellipse matrix and changes in RSS
Beginning the implementation of the RF model: adding the functions
locate_ellipses and rss_links.
The first models and ellipse between each link (two sensors) and gives a
weight to each voxel inside the ellipse (matrix W). This data is fixed
for all the steps.
The function rss_links calculates the changes in the RSS (received
signal strength) in each link (y = W*x+n). This data changes for each
step because depends on the initial radiation. The vector "x" contents
the changes caused by each voxel to the RSS. In this point is equal to
the gaussian radiation caused by each target. To model better the RF
behaviour some changes must be add (Rayleigh fading, etc).
The gaussianradiation function has also been changed in order to reduce
the execution time.
Also other minor changes in the rest of functions.